fix: Move sorting of paths to helper file and small tweak to empty descriptions in comments
