+++ test/resources/examples/portal/src/mocks/handlers/tasks/addTaskHandler.ts
@@ -2,6 +2,7 @@ import { http, passthrough } from 'msw';
 
 /**
  * Add task for user
+ *
  * HTTP: POST /api/rest/tasks/tasks/:userId/:id
  * Response type: Task
  */
