+++ test/resources/examples/portal/src/mocks/handlers/users/removeUserHandler.ts
@@ -2,6 +2,7 @@ import { http, HttpResponse } from 'msw';
 
 /**
  * Delete user
+ *
  * HTTP: DELETE /api/rest/users/users/:id
  * Response type: void
  */
